Moved engines to lib
