+++ lib/text2speech.js
@@ -2,7 +2,7 @@
 
 function Text2Speech(adapter, libs, options, sayIt) {
     const that         = this;
-    const engines      = require('../admin/engines.js');
+    const engines      = require('../lib/engines.js');
     const googleTTS    = require('google-tts-api');
     const sayitEngines = engines.sayitEngines;
     const MP3FILE      = `${adapter.config.dataDir}/${adapter.namespace}.say.${options.outFileExt}`;
